// frontend/lib/api/generated/timeAnchorResponse.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */

/**
 * 时间锚点响应
 */
export interface TimeAnchorResponse {
  /** ID */
  id: number;
  /** 会议ID */
  conferenceId?: string;
  /** 更新时间 */
  updateTime?: string;
  /** 锚点物理时间 */
  anchorRealTime?: string;
  /** 锚点游戏时间 */
  anchorGameTime?: string;
  /** 流速 */
  timeRatio?: number;
  /** 是否当前生效 */
  isCurrent: boolean;
}